# crates/api-client/openapi.gen.json
{
  "components": {
    "schemas": {
...
      "ApiKeyInfo": {
        "properties": {
          "created_at": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "key_prefix": {
            "type": "string"
          },
          "last_used_at": {
            "nullable": true,
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "key_prefix",
          "created_at"
        ],
        "type": "object"
      },
...
      "CloudApiSettings": {
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "updated_at": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "enabled"
        ],
        "type": "object"
      },
...
      "CreateApiKeyBody": {
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
...
      "CreatedApiKey": {
        "properties": {
          "created_at": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "key_prefix": {
            "type": "string"
          },
          "last_used_at": {
            "nullable": true,
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "key_prefix",
          "created_at",
          "key"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "bearerFormat": "JWT",
        "description": "Supabase JWT token",
        "scheme": "bearer",
        "type": "http"
      },
      "cloud_api_key": {
        "bearerFormat": "anl_...",
        "description": "Session Echo cloud API key",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Session Echo cloud services and opt-in hosted meeting access",
    "license": {
      "name": ""
    },
    "title": "Session Echo API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/v1/cloud-api/keys": {
      "get": {
        "operationId": "list_keys",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ApiKeyInfo"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          },
          "403": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "tags": [
          "cloud-api"
        ]
      },
      "post": {
        "operationId": "create_key",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApiKeyBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedApiKey"
                }
              }
            },
            "description": ""
          },
          "400": {
            "description": ""
          },
          "403": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "tags": [
          "cloud-api"
        ]
      }
    },
    "/v1/cloud-api/keys/{key_id}": {
      "delete": {
        "operationId": "revoke_key",
        "parameters": [
          {
            "description": "Cloud API key id",
            "in": "path",
            "name": "key_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "tags": [
          "cloud-api"
        ]
      }
    },
    "/v1/cloud-api/settings": {
      "get": {
        "operationId": "get_settings",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CloudApiSettings"
                }
              }
            },
            "description": ""
          },
          "403": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "tags": [
          "cloud-api"
        ]
      },
      "put": {
        "operationId": "update_settings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSettingsBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CloudApiSettings"
                }
              }
            },
            "description": ""
          },
          "403": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "tags": [
          "cloud-api"
        ]
      }
    },
    "/v1/meetings": {
      "get": {
        "operationId": "list_meetings",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "series_id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingPage"
                }
              }
            },
            "description": ""
          },
          "401": {
            "description": ""
          },
          "403": {
            "description": ""
          }
        },
        "security": [
          {
            "cloud_api_key": []
          }
        ],
        "tags": [
          "cloud-api"
        ]
      }
    },
    "/v1/meetings/{meeting_id}": {
      "get": {
        "operationId": "get_meeting",
        "parameters": [
          {
            "in": "path",
            "name": "meeting_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meeting"
                }
              }
            },
            "description": ""
          },
          "404": {
            "description": ""
          }
        },
        "security": [
          {
            "cloud_api_key": []
          }
        ],
        "tags": [
          "cloud-api"
        ]
      }
    },
    "/v1/meetings/{meeting_id}/export": {
      "get": {
        "operationId": "export_meeting",
        "parameters": [
          {
            "in": "path",
            "name": "meeting_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "json or markdown",
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Meeting export"
          },
          "400": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        },
        "security": [
          {
            "cloud_api_key": []
          }
        ],
        "tags": [
          "cloud-api"
        ]
      }
    },
    "/v1/meetings/{meeting_id}/history": {
      "get": {
        "operationId": "get_history",
        "parameters": [
          {
            "in": "path",
            "name": "meeting_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingPage"
                }
              }
            },
            "description": ""
          },
          "404": {
            "description": ""
          }
        },
        "security": [
          {
            "cloud_api_key": []
          }
        ],
        "tags": [
          "cloud-api"
        ]
      }
    },
    "/v1/meetings/{meeting_id}/transcript": {
      "get": {
        "operationId": "get_transcript",
        "parameters": [
          {
            "in": "path",
            "name": "meeting_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscriptPage"
                }
              }
            },
            "description": ""
          },
          "404": {
            "description": ""
          }
        },
        "security": [
          {
            "cloud_api_key": []
          }
        ],
        "tags": [
          "cloud-api"
        ]
      }
    },
...
  },
...
}